chore(gardener): details dialog
